refactor: exception handling

+++ set_config.sh
@@ -1,6 +1,8 @@
 #! /bin/sh
 
-rm ~/.gitconfig
+if [ -f ~/.gitconfig ]; then
+  rm ~/.gitconfig
+fi
 ln -s ~/.git/gitconfig ~/.gitconfig
 
 echo "Set Git Configuration!"
